# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/customers/{id}" : {
      "put" : {
        "tags" : [ "customer-profile-controller" ],
        "operationId" : "updateProfile",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Customer"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCustomer"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/orders/{orderId}/partial-refund" : {
      "post" : {
        "tags" : [ "internal-order-refund-controller" ],
        "operationId" : "initiatePartialRefund",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefundCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RefundView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/restaurants/nearby" : {
      "get" : {
        "tags" : [ "customer-restaurant-controller" ],
        "operationId" : "getNearbyRestaurants",
        "parameters" : [ {
          "name" : "lat",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        }, {
          "name" : "lng",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        }, {
          "name" : "radius",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "number",
            "format" : "double",
            "default" : 5.0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/brands/{brandId}/outlets" : {
      "get" : {
        "tags" : [ "customer-restaurant-controller" ],
        "operationId" : "getBrandOutlets",
        "parameters" : [ {
          "name" : "brandId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "lat",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        }, {
          "name" : "lng",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        }, {
          "name" : "radius",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "number",
            "format" : "double",
            "default" : 5.0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/places/reverse-geocode" : {
      "get" : {
        "tags" : [ "places-controller" ],
        "operationId" : "reverseGeocode",
        "parameters" : [ {
          "name" : "lat",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        }, {
          "name" : "lng",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMapStringObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/places/autocomplete" : {
      "get" : {
        "tags" : [ "places-controller" ],
        "operationId" : "autocomplete",
        "parameters" : [ {
          "name" : "input",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListMapStringObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Customer" : {
        "required" : [ "createdAt", "id", "phoneNumber" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseCustomer" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/Customer"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "Item" : {
        "type" : "object",
        "properties" : {
          "orderItemId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "RefundCommand" : {
        "type" : "object",
        "properties" : {
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "type" : "number"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Item"
            }
          },
          "reasonCode" : {
            "type" : "string"
          },
          "reasonText" : {
            "type" : "string"
          },
          "faultType" : {
            "type" : "string",
            "enum" : [ "PLATFORM_FAULT", "RESTAURANT_FAULT", "RIDER_FAULT", "CUSTOMER_FAULT", "UNKNOWN" ]
          },
          "destination" : {
            "type" : "string",
            "enum" : [ "ORIGINAL_METHOD", "STORE_CREDIT", "NONE" ]
          },
          "source" : {
            "type" : "string",
            "enum" : [ "CUSTOMER_TICKET", "RESTAURANT", "ADMIN", "SYSTEM_CANCELLATION", "SYSTEM_DELIVERY_FAILED", "SYSTEM_LATE_PAYMENT" ]
          },
          "initiatorType" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "RESTAURANT", "ADMIN", "SYSTEM" ]
          },
          "initiatorId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "ticketId" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
      "RefundView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "type" : "number"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "REQUESTED", "PROCESSING", "COMPLETED", "FAILED", "CANCELLED" ]
          },
          "destination" : {
            "type" : "string",
            "enum" : [ "ORIGINAL_METHOD", "STORE_CREDIT", "NONE" ]
          },
          "method" : {
            "type" : "string",
            "enum" : [ "CARD", "UPI", "WALLET", "COD" ]
          },
          "reasonCode" : {
            "type" : "string"
          },
          "requestedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "completedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expectedBy" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ApiResponseMapStringObject" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ApiResponseListObject" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "type" : "object"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseListMapStringObject" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}